services/ingestion/requirements.txt
anthropic==0.39.0
fastapi==0.110.3
httpx==0.27.2
pydantic==2.9.2
PyJWT==2.9.0
PyMuPDF==1.24.10
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
redis==5.0.8
uvicorn==0.30.6
